Update Home.html
I have my home page, which has a hero image and some simple info. I am working on adding a contact page that will have my contact info. I am working on smoothing out some things. One thing I am having trouble with, is for some reason, it says that there is more "content to the right." I have tried commenting out certain parts to try and figure out what is causing it so I could solve the problem, but no luck so far. Also, I forgot that I have to use the git desktop to commit to github. So I was working on it thinking that it was just saving to github, but it wasn't. So here I am committing my work to github now.